restructure folder content add SignInPage
