+++ src/App.js
@@ -1,16 +1,16 @@
-import React, { Component } from 'react';
-import './App.css';
-import SignInCard from './components/SignInCard';
-import { Page } from './DefaultComponents'
+import React, { useState } from 'react'
+import './App.css'
+import SignInPage from './pages/SignInPage';
 
-class App extends Component {
-  render () {
-    return (
-      <Page>
-        <SignInCard />
-      </Page>
-    )
-  }
+function App () {
+  const [signedIn, setSignedIn] = useState(false)
+
+  return (
+    <div>
+      {signedIn ? <h1>Ready To Go</h1> : <SignInPage></SignInPage>}
+    </div>
+
+  )
 }
 
 export default App
