+++ src/pages/SignInPage.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Page } from '../DefaultComponents';
+import SignInCard from '../components/SignInCard';
+
+function SignInPage () {
+  return (
+    <Page>
+      <SignInCard></SignInCard>
+    </Page>
+  )
+}
+
+export default SignInPage
